% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_function.R
\name{check_function}
\alias{check_function}
\title{check_function}
\usage{
check_function(x)
}
\arguments{
\item{data}{that needs to be checked}
}
\value{
a list
}
\description{
check_function
}
\examples{
test1 <- c("a", "b", "c")
check_function(test1)
}
