% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_type.R
\name{check_type}
\alias{check_type}
\title{check_type}
\usage{
check_type(x)
}
\arguments{
\item{data}{that needs to be checked for type}
}
\value{
a list
}
\description{
check_type
}
\examples{
test1 <- c("a", "b", "c")
check_type(test1)
}
