% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{Calculating mean and standard deviation}
\usage{
mean_sd(df)
}
\arguments{
\item{df}{data frame for calculating the mean and the standard deviation}
}
\value{
a list
}
\description{
Calculating mean and standard deviation
}
\examples{
test4 <- c("1", "2", "3")
mean_sd(test4)
}
