% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mima.R
\name{mima}
\alias{mima}
\title{Calculating minimum and maximum value}
\usage{
mima(df)
}
\arguments{
\item{df}{data frame for calculating the minimum and maximum}
}
\value{
a list
}
\description{
Calculating minimum and maximum value
}
\examples{
name <- c("Noa", "Luke", "Oliver", "Anna", "Julie")
gender <- c("F", "M", "M", "F", "F")
age <- c(21, 29, 27, 29, 24)
test <- data.frame(name, gender, age)
mima(test$age)
}
