% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_check.R
\name{multi_check}
\alias{multi_check}
\title{Class check of multiple datasets}
\usage{
multi_check(x, y, z)
}
\arguments{
\item{x}{data$column of first set, column can be left out}

\item{y}{data$column of second set, column can be left out}

\item{z}{data$column of third set, column can be left out}
}
\value{
a list
}
\description{
Class check of multiple datasets
}
\examples{
name <- c("Noa", "Luke", "Oliver", "Anna", "Julie")
gender <- c("F", "M", "M", "F", "F")
age <- c(21, 29, 27, 29, 24)
test <- data.frame(name, gender, age)
multi_check(test$name, test$gender, test$age)
}
