% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_function.R
\name{save_function}
\alias{save_function}
\title{save_function}
\usage{
save_function(df, output_name, rownames = FALSE)
}
\arguments{
\item{df}{dataframe to save as CSV and RSD}

\item{output_name}{name of output file and location}

\item{rownames}{rownames TRUE or FALSE}
}
\value{
CSV and RSD
}
\description{
save_function
}
\examples{
test2 <- data.frame(col1 = c("a", "b", "c"), col2 = c("d", "e", "f"))
save_function(test2, "test2")
}
