% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{Calculating mean and standard deviation}
\usage{
mean_sd(df)
}
\arguments{
\item{df}{data frame for calculating the mean and the standard deviation}
}
\value{
a list
}
\description{
Calculating mean and standard deviation
}
\examples{
name <- c("Noa", "Luke", "Oliver", "Anna", "Julie")
gender <- c("F", "M", "M", "F", "F")
age <- c(21, 29, 27, 29, 24)
test <- data.frame(name, gender, age)
mean_sd(test$age)
}
