% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mima.R
\name{mima}
\alias{mima}
\title{Calculating minimum and maximum value}
\usage{
mima(df)
}
\arguments{
\item{df}{data frame for calculating the minimum and maximum}
}
\value{
a list
}
\description{
Calculating minimum and maximum value
}
\examples{
test3 <- c("1", "2", "3", "4")
mima(test3)
}
