% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_check.R
\name{multi_check}
\alias{multi_check}
\title{Class check of multiple datasets}
\usage{
multi_check(x, y, z)
}
\arguments{
\item{x}{data$column of first set, column can be left out}

\item{y}{data$column of second set, column can be left out}

\item{z}{data$column of third set, column can be left out}
}
\value{
a list
}
\description{
Class check of multiple datasets
}
\examples{
test3a <- data.frame(col1 = c("1", "2", "3"), col2 = c("a", "b", "c"))
test3b <- c("4", "5", "6")
multi_check(test3a$col1, test3a$col2, test3b)
}
