% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_csv_rds.R
\name{save_to_csv_rds}
\alias{save_to_csv_rds}
\title{save_to_csv_rds}
\usage{
save_to_csv_rds(df, output_name, rownames = FALSE)
}
\arguments{
\item{df}{dataframe to save as CSV and RDS}

\item{output_name}{name of output file and location}

\item{rownames}{rownames TRUE or FALSE}
}
\value{
CSV and RDS
}
\description{
save_to_csv_rds
}
\examples{
test2 <- data.frame(col1 = c("a", "b", "c"), col2 = c("d", "e", "f"))
save_to_csv_rds(test2, "test2")
}
