% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_csv_rds.R
\name{save_to_csv_rds}
\alias{save_to_csv_rds}
\title{Save data to both CSV and RDS files}
\usage{
save_to_csv_rds(df, output_name, rownames = FALSE)
}
\arguments{
\item{df}{dataframe to save as CSV and RDS}

\item{output_name}{name of output file and location}

\item{rownames}{rownames TRUE or FALSE}
}
\value{
CSV and RDS
}
\description{
Save data to both CSV and RDS files
}
\examples{
name <- c("Noa", "Luke", "Oliver", "Anna", "Julie")
gender <- c("F", "M", "M", "F", "F")
age <- c(21, 29, 27, 29, 24)
test <- data.frame(name, gender, age)
save_to_csv_rds(test, "test")
}
